reorganize meta tag index into distinct files
